local-alpha support
If the V4L2 device supports it, use alpha blending instead of colorkey.  This
requires that the framebuffer is in RGBA mode.  And some trick is required to
ensure that the alpha channel value for entire UI (except where the video is)
has value of 0xff.

In addition to the earlier patch on omap_vout driver for xoverlay support,
the following change is needed in omapfb-main.c:

---------------
@@ -360,7 +362,8 @@ static int fb_mode_to_dss_mode(struct fb_var_screeninfo *var,
                dssmode = OMAP_DSS_COLOR_RGB24P;
                break;
        case 32:
-               dssmode = OMAP_DSS_COLOR_RGB24U;
+               dssmode = OMAP_DSS_COLOR_ARGB32;
                break;
        default:
                return -EINVAL;
--------------

Probably xf86-video-v4l2 driver should gain some support to reconfigure the
appropriate framebuffer device so we don't need the above kernel hack.. but
I need to figure out how to do this.